Tutorial-4/deployment/python_app/requirements.txt
flask
paho-mqtt
apscheduler